requirements.txt
playwright==1.50.0
python-dotenv==1.0.1
openai==1.61.0
requests==2.32.3
httpx>=0.27.0
